Fix test and add further explanations
Due to the refactor done in a previous commit, one of the tests now
fails.

This commit fixes that, changing also the approach taken in the
refactored code.

This issue showcases the problem of embedding business logic into code
that depends on the servlet API. It might be a very interesting showcase
to explain to the whole class.